Fix dtype error in compute_group_observations
- Use torch.float32 instead of self.obs_buf.dtype for heightmap_data
- Prevents TypeError when obs_buf is not yet initialized

+++ legged_gym/envs/pointfoot_flat/pointfoot_flat.py
@@ -315,7 +315,7 @@ class BipedPF(BaseTask):
             heightmap_data = self.measured_heights
         else:
             # Fallback to zeros if no heightmap available
-            heightmap_data = torch.zeros((self.num_envs, 81), device=self.device, dtype=self.obs_buf.dtype)
+            heightmap_data = torch.zeros((self.num_envs, 81), device=self.device, dtype=torch.float32)
         
         critic_obs_buf = torch.cat((
             self.base_lin_vel * self.obs_scales.lin_vel, self.obs_buf, heightmap_data), dim=-1)
